requirements.txt
streamlit
google-genai
pypdf
python-docx
EbookLib
beautifulsoup4
pyvis
networkx
